chg: Map renamed to AMaps to correct the naming confilct in the unittests
