+++ app/components/AMaps/index.test.js
